Move server code to flwr.server
