+++ src/py/flwr/__init__.py
@@ -15,7 +15,4 @@
 """Flower main package."""
 
 
-from . import app, client
-from .client_manager import SimpleClientManager
-from .history import History
-from .server import Server
+from . import client, server
