+++ src/py/flwr/server/app.py
@@ -18,11 +18,11 @@
 from logging import INFO
 from typing import Dict, Optional
 
-from flwr.client_manager import SimpleClientManager
 from flwr.common.logger import log
-from flwr.grpc_server.grpc_server import start_insecure_grpc_server
-from flwr.server import Server
-from flwr.strategy import FedAvg, Strategy
+from flwr.server.client_manager import SimpleClientManager
+from flwr.server.grpc_server.grpc_server import start_insecure_grpc_server
+from flwr.server.server import Server
+from flwr.server.strategy import FedAvg, Strategy
 
 DEFAULT_SERVER_ADDRESS = "[::]:8080"
