+++ src/py/flwr/server/strategy/fast_and_slow_test.py
@@ -16,7 +16,7 @@
 
 import unittest
 
-from flwr.strategy import fast_and_slow
+from flwr.server.strategy import fast_and_slow
 
 
 # pylint: disable-msg=no-self-use,missing-class-docstring,missing-function-docstring
