+++ src/py/flwr/server/strategy/fault_tolerant_fedavg_test.py
@@ -18,8 +18,8 @@
 from typing import List, Optional, Tuple
 from unittest.mock import MagicMock
 
-from flwr.client_proxy import ClientProxy
 from flwr.common import EvaluateRes, FitRes, Parameters, Weights
+from flwr.server.client_proxy import ClientProxy
 
 from .fault_tolerant_fedavg import FaultTolerantFedAvg
