+++ src/py/flwr_example/quickstart/server.py
@@ -1,4 +1,4 @@
 import flwr as fl
 
 if __name__ == "__main__":
-    fl.app.server.start_server(config={"num_rounds": 3})
+    fl.server.start_server(config={"num_rounds": 3})
